Add the Corstone1000 platform. On this platform the PSA RoT runs on a
dedicated Cortex-M subsystem called Secure Enclave (SE), and the
Cortex-A subsystem (AP) acts as the Non-secure Processing Environment.
In this setup all PSA Services on the AP are mapped to a single proxy-sp
deployment. This allows making PSA requests from the AP to the PSA RoT.

This merge:
  - adds new components which implement the TF-M specific psa-ipc. This
    allows making PSA calls to TF-M Secure Side from Cortex-A cores.
  - adds the new corstone1000 platform with an MHUv2 driver. The Message
    Handling Uint is the HW peripheral trough which the SE and the AP
	subsystem communicates.
  - adds OpenAMP and its porting layer libmetal as external components
  - adds a new configuration to the proxy-sp deployment to allow
    building the deployment for Corstone1000
  - updates the UEFI Variable service to:
      - use the Proxy SP for storage on the new platform
	  - use a proper variable size
	  - treat null sized set operations as delete requests

+++ components/service/secure_storage/include/psa/storage_common.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, Arm Limited and Contributors. All rights reserved.
+ * Copyright (c) 2020-2023, Arm Limited and Contributors. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
@@ -20,8 +20,8 @@ typedef uint64_t psa_storage_uid_t;
 typedef uint32_t psa_storage_create_flags_t;
 
 struct psa_storage_info_t {
-	size_t capacity;
-	size_t size;
+	uint32_t capacity;
+	uint32_t size;
 	psa_storage_create_flags_t flags;
 };
